Clarify SwapProtocol registry type
The abstraction between a swap protocol and the more general COMIT protocol is
currently easy to blur because we only have swap protocols (just the one,
rfc-003).  No protocols other than swap protocols are going to be added in the
foreseeable future so there is no need to make this abstraction difference
explicit at the moment.  For now just clarify what we are referring to.