Document 32 byte secret length limit
Recently an bug was found allowing an attack against COMIT if the secret used
was longer than 32 bytes AND the alpha/beta ledgers were Ethereum/Bitcoin.  This
is now patched.  We now only consider the first 32 bytes of the secret, this
should be documented in the RFC since secret length is already discussed.

Update discussion of the secret length to explicitly state that length used is
only 32 bytes.