Link terms sender/receiver with alice/bob
We define the 'sender' and 'receiver' then much further down the document link
them with the terms 'alice' and 'bob'.  We should do this linking on the first
usage of sender/receiver.  This aids clarity and is typical in written documents
e.g. defining an acronym on first usage of the term.

Explicitly link the terms sender/receiver with alice/bob respectively.