Remove ERROR frame from rfc-001
Recent work attempting to implement the ERROR frame as described in
rcf-001 found that the ERROR frame was not the best solution to the
problem it was trying to solve. For now we decided to just close the
stream for message level errors.  We should update the RFC to mirror
what was implemented in `comit-rs`.  Since rcf-001 does not yet mention
libp2p; specifically do not mention the word 'stream' instead use
'communication channel'.

Remove ERROR frame from rfc-001.  Add description of current error
handling behaviour implemented in `comit-rs`.

Fixes: #117